+++ CSVReader.h
@@ -21,7 +21,7 @@ class CSVReader
                                         std::string product, 
                                         OrderBookType OrderBookType);
     
-   std::map<std::string, std::map<std::string, std::vector<double>>> readTemperatureData(const std::string& filename);
+    std::map<std::string, std::map<std::string, std::vector<float>>> readTemperatureData(const std::string& filename);
    // std::vector<std::string> readDates(const std::string& filename, const std::string& country);
     private:
      static OrderBookEntry stringsToOBE(std::vector<std::string> strings);
